split, move & rename
