+++ ui/submodules/packages/components/src/common/hooks/index.ts
@@ -1,0 +1,1 @@
+export { useScrollParent } from "./useScrollParent";
