+++ ui/submodules/packages/components/src/scenarios/filters/index.ts
@@ -1,0 +1,2 @@
+export { filterRules } from "./filterRules";
+export { FiltersRow } from "./filtersRow";
