+++ ui/submodules/packages/components/src/scenarios/rootRoutes.tsx
@@ -1,13 +1,13 @@
 import React from "react";
 import { Route, Routes } from "react-router-dom";
 import { UnavailableViewPlaceholder, View } from "../common";
-import { Prototype } from "./prototype";
+import { ScenariosView } from "./scenariosView";
 
 export function RootRoutes({ inTab }: { inTab?: boolean }): JSX.Element {
     return (
         <View inTab={inTab}>
             <Routes>
-                <Route path="/" element={<Prototype />} />
+                <Route path="/" element={<ScenariosView />} />
                 <Route path="*" element={<UnavailableViewPlaceholder />} />
             </Routes>
         </View>
